safari-600.7-branch/LayoutTests/fast/events/touch/gesture/touch-gesture-scroll-sideways-expected.txt

This tests that precise scrolls on a horizontal scroll bar move vertically.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


first wheel event
wheel event 0+> [object HTMLDivElement]
PASS movingdiv.scrollTop is 19
PASS movingdiv.scrollLeft is 0
PASS wheelEventsOccurred is 1
second wheel event
wheel event 1+> [object HTMLDivElement]
without precise deltas, scrollLeft: 0 scrollTop: 44
PASS wheelEventsOccurred is 2
scroll event 0+> [object HTMLDivElement]
PASS scrollEventsOccurred is 1
PASS successfullyParsed is true

TEST COMPLETE
